+++ InstaConnect.Business/Services/PostCommentLikeService.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
 using InstaConnect.Business.Abstraction.Factories;
 using InstaConnect.Business.Abstraction.Services;
+using InstaConnect.Business.Models.DTOs.Message;
 using InstaConnect.Business.Models.DTOs.PostCommentLike;
 using InstaConnect.Business.Models.Results;
 using InstaConnect.Business.Models.Utilities;
+using InstaConnect.Data.Abstraction.Helpers;
 using InstaConnect.Data.Abstraction.Repositories;
 using InstaConnect.Data.Models.Entities;
 
@@ -16,19 +18,22 @@ namespace InstaConnect.Business.Services
         private readonly IPostCommentLikeRepository _postCommentLikeRepository;
         private readonly IPostCommentRepository _commentRepository;
         private readonly IUserRepository _userRepository;
+        private readonly IAccountManager _accountManager;
 
         public PostCommentLikeService(
             IMapper mapper,
             IResultFactory resultFactory,
             IPostCommentLikeRepository postCommentLikeRepository,
             IPostCommentRepository commentRepository,
-            IUserRepository userRepository)
+            IUserRepository userRepository,
+            IAccountManager accountManager)
         {
             _mapper = mapper;
             _resultFactory = resultFactory;
             _postCommentLikeRepository = postCommentLikeRepository;
             _commentRepository = commentRepository; ;
             _userRepository = userRepository;
+            _accountManager = accountManager;
         }
 
         public async Task<IResult<ICollection<PostCommentLikeResultDTO>>> GetAllAsync(
@@ -83,8 +88,17 @@ namespace InstaConnect.Business.Services
             return okResult;
         }
 
-        public async Task<IResult<PostCommentLikeResultDTO>> AddAsync(PostCommentLikeAddDTO postCommentLikeAddDTO)
+        public async Task<IResult<PostCommentLikeResultDTO>> AddAsync(string userId, PostCommentLikeAddDTO postCommentLikeAddDTO)
         {
+            var validUser = _accountManager.ValidateUser(userId, postCommentLikeAddDTO.UserId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<PostCommentLikeResultDTO>();
+
+                return forbiddenResult;
+            }
+
             var existingUser = await _userRepository.FindEntityAsync(f => f.Id == postCommentLikeAddDTO.UserId);
 
             if (existingUser == null)
@@ -140,13 +154,22 @@ namespace InstaConnect.Business.Services
 
         public async Task<IResult<PostCommentLikeResultDTO>> DeleteAsync(string userId, string id)
         {
-            var existingPostCommentLike = await _postCommentLikeRepository.FindEntityAsync(cl => cl.Id == id && cl.UserId == userId);
+            var existingPostCommentLike = await _postCommentLikeRepository.FindEntityAsync(cl => cl.Id == id);
 
             if (existingPostCommentLike == null)
             {
                 var notFoundResult = _resultFactory.GetNotFoundResult<PostCommentLikeResultDTO>();
 
                 return notFoundResult;
+            }
+
+            var validUser = _accountManager.ValidateUser(userId, existingPostCommentLike.UserId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<PostCommentLikeResultDTO>();
+
+                return forbiddenResult;
             }
 
             await _postCommentLikeRepository.DeleteAsync(existingPostCommentLike);
