+++ InstaConnect.Business/Services/PostService.cs
@@ -2,8 +2,10 @@
 using InstaConnect.Business.Abstraction.Factories;
 using InstaConnect.Business.Abstraction.Services;
 using InstaConnect.Business.Models.DTOs.Post;
+using InstaConnect.Business.Models.DTOs.PostLike;
 using InstaConnect.Business.Models.Results;
 using InstaConnect.Business.Models.Utilities;
+using InstaConnect.Data.Abstraction.Helpers;
 using InstaConnect.Data.Abstraction.Repositories;
 using InstaConnect.Data.Models.Entities;
 
@@ -15,17 +17,20 @@ namespace InstaConnect.Business.Services
         private readonly IResultFactory _resultFactory;
         private readonly IPostRepository _postRepository;
         private readonly IUserRepository _userRepository;
+        private readonly IAccountManager _accountManager;
 
         public PostService(
             IMapper mapper,
             IResultFactory resultFactory,
             IPostRepository postRepository,
-            IUserRepository userRepository)
+            IUserRepository userRepository,
+            IAccountManager accountManager)
         {
             _mapper = mapper;
             _resultFactory = resultFactory;
             _postRepository = postRepository;
             _userRepository = userRepository;
+            _accountManager = accountManager;
         }
 
         public async Task<IResult<ICollection<PostResultDTO>>> GetAllAsync(
@@ -63,8 +68,17 @@ namespace InstaConnect.Business.Services
             return okResult;
         }
 
-        public async Task<IResult<PostResultDTO>> AddAsync(PostAddDTO postAddDTO)
+        public async Task<IResult<PostResultDTO>> AddAsync(string userId, PostAddDTO postAddDTO)
         {
+            var validUser = _accountManager.ValidateUser(userId, postAddDTO.UserId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<PostResultDTO>();
+
+                return forbiddenResult;
+            }
+
             var existingUser = await _userRepository.FindEntityAsync(f => f.Id == postAddDTO.UserId);
 
             if (existingUser == null)
@@ -84,13 +98,22 @@ namespace InstaConnect.Business.Services
 
         public async Task<IResult<PostResultDTO>> UpdateAsync(string userId, string id, PostUpdateDTO postUpdateDTO)
         {
-            var existingPost = await _postRepository.FindEntityAsync(p => p.Id == id && p.UserId == userId);
+            var existingPost = await _postRepository.FindEntityAsync(p => p.Id == id);
 
             if (existingPost == null)
             {
                 var notFoundResult = _resultFactory.GetNotFoundResult<PostResultDTO>();
 
                 return notFoundResult;
+            }
+
+            var validUser = _accountManager.ValidateUser(userId, existingPost.UserId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<PostResultDTO>();
+
+                return forbiddenResult;
             }
 
             _mapper.Map(postUpdateDTO, existingPost);
@@ -103,13 +126,22 @@ namespace InstaConnect.Business.Services
 
         public async Task<IResult<PostResultDTO>> DeleteAsync(string userId, string id)
         {
-            var existingPost = await _postRepository.FindEntityAsync(p => p.Id == id && p.UserId == userId);
+            var existingPost = await _postRepository.FindEntityAsync(p => p.Id == id);
 
             if (existingPost == null)
             {
                 var notFoundResult = _resultFactory.GetNotFoundResult<PostResultDTO>();
 
                 return notFoundResult;
+            }
+
+            var validUser = _accountManager.ValidateUser(userId, existingPost.UserId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<PostResultDTO>();
+
+                return forbiddenResult;
             }
 
             await _postRepository.DeleteAsync(existingPost);
